メモがない状態で-d,-rを実行するとエラーになるのを修正、id順で取得するように

+++ 04.class/memo_app.js
@@ -34,11 +34,17 @@ async function main() {
       });
     }
     if (argv.r) {
+      if (memos.length === 0){
+        return console.log("メモはまだありません")
+      }
       const answer = await selectMemo(memos);
       const found = memos.find((memo) => memo.id === answer);
       console.log(found.content);
     }
     if (argv.d) {
+      if (memos.length === 0){
+        return console.log("メモはまだありません")
+      }
       const answer = await selectMemo(memos);
       await db.deleteMemo(answer);
     }
